dexter_vietnam/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.24.0

# LLM Integration
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=0.0.6

# Data & Analysis
vnstock>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
ta>=0.11.0  # Technical analysis library

# Web scraping
selenium>=4.16.0
playwright>=1.40.0

# API & Web
fastapi>=0.109.0
uvicorn>=0.27.0

# CLI
rich>=13.7.0
click>=8.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
